renameall update Renameall now builds a database in memory, for faster word search and lookup, instead of the old method of going through the dictionary file randomly, until a short enough word is found.
I decided to make it in memory and not persistent since it's fast and easy to generate, and I'm trying to keep storage requirements to a minimum, as to not go over [replit.com](https://replit.com)'s storage limit.

It adds about 1 second to the startup time, and considering it runs for hours at a time, this doesn't seem like an issue.

The new way is also easier to extend with new formatting options.